juju/client: factor out JujuData class
We move JujuData into its own file because it's logically
independent of the Connection class.

We also make it cache the data that's been read so that we don't need
to read and parse the YAML file every time a method
is called.

We also change it so that it doesn't shell out to Juju to obtain the
current controller name - the name can be retrieved from the controllers.yaml
file.

In a subsequent PR, we'll factor those Connection methods
that depend on JujuData out of the Connection class so
that Connection doesn't need to depend on the details of
the juju configuration files.
